+++ src/resolvers/validate-arg.ts
@@ -9,13 +9,13 @@ import { ValidatorFn } from "../interfaces/ValidatorFn";
 const shouldArgBeValidated = (argValue: unknown): boolean =>
   argValue !== null && typeof argValue === "object";
 
-export async function validateArg<T extends object>(
-  argValue: T | undefined,
+export async function validateArg(
+  argValue: any | undefined,
   argType: TypeValue,
   globalValidate: ValidateSettings,
   argValidate: ValidateSettings | undefined,
-  validateFn: ValidatorFn<object> | undefined,
-): Promise<T | undefined> {
+  validateFn: ValidatorFn | undefined,
+): Promise<any | undefined> {
   if (typeof validateFn === "function") {
     await validateFn(argValue, argType);
     return argValue;
@@ -48,7 +48,7 @@ export async function validateArg<T extends object>(
           .map(argItem => validateOrReject(argItem, validatorOptions)),
       );
     } else {
-      await validateOrReject(argValue as T, validatorOptions);
+      await validateOrReject(argValue, validatorOptions);
     }
     return argValue;
   } catch (err) {